Rarranged payment index functions
